modify variable byte width to uint_32

+++ helper/priority-port-tag.h
@@ -1,10 +1,8 @@
-// PriorityPortTag.h
 #ifndef PRIORITY_PORT_TAG_H
 #define PRIORITY_PORT_TAG_H
 
 #include "ns3/tag.h"
 #include "ns3/nstime.h"
-#include "ns3/uinteger.h"
 
 namespace ns3 {
 
@@ -15,13 +13,13 @@ public:
   TypeId GetInstanceTypeId (void) const override;
 
   PriorityPortTag ();
-  PriorityPortTag (uint8_t priority, uint16_t port);
+  PriorityPortTag (uint32_t priority, uint32_t port);
 
-  void SetPriority (uint8_t priority);
-  uint8_t GetPriority (void) const;
+  void SetPriority (uint32_t priority);
+  uint32_t GetPriority (void) const;
 
-  void SetPort (uint16_t port);
-  uint16_t GetPort (void) const;
+  void SetPort (uint32_t port);
+  uint32_t GetPort (void) const;
 
   void Serialize (TagBuffer i) const override;
   void Deserialize (TagBuffer i) override;
@@ -30,8 +28,8 @@ public:
   void Print (std::ostream &os) const override;
 
 private:
-  uint8_t m_priority;
-  uint16_t m_port;
+  uint32_t m_priority;
+  uint32_t m_port;
 };
 
 } // namespace ns3
